change names and bugfix

+++ RecoBTag/CTagging/python/charmTaggerProducer_cfi.py
@@ -4,7 +4,7 @@ charmTagsComputer = cms.ESProducer(
    'CharmTaggerESProducer',
    weightFile = cms.FileInPath('RecoBTag/CTagging/data/c_vs_udsg.weight.xml'),
    variables = cms.VPSet(
-      cms.PSet(name = cms.string('dummy'), default = cms.float(0))
+      cms.PSet(name = cms.string('dummy'), default = cms.double(0))
       ),
    computer = cms.ESInputTag('combinedSecondaryVertexSoftLeptonComputer'),
    tagInfos = cms.VInputTag(
